✏️[BE]Budget 코드 수정

+++ showmethemoney2/src/main/java/com/example/showmethemoney2/entity/Budget.java
@@ -53,9 +53,9 @@ import lombok.Setter;
         private String username;
         private int year;
         private int month;
-        private double budget;
+        private int budget;
 
-        public Budget(String username, int year, int month, double budget) {
+        public Budget(String username, int year, int month, int budget) {
             this.username = username;
             this.year = year;
             this.month = month;
